+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/validation/Validations.scala
@@ -5,12 +5,13 @@ import cats.data.ValidatedNel
 import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.engine.api.context.PartSubGraphCompilationError
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.InvalidVariableName
+import pl.touk.nussknacker.engine.api.parameter.ParameterName
 
 import javax.lang.model.SourceVersion
 
 object Validations {
 
-  def validateVariableName(name: String, paramName: Option[String])(
+  def validateVariableName(name: String, paramName: Option[ParameterName])(
       implicit nodeId: NodeId
   ): ValidatedNel[PartSubGraphCompilationError, String] = {
     // TODO: add correct and more precise error messages