+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/deployment/EngineSetupName.scala
@@ -1,6 +1,8 @@
 package pl.touk.nussknacker.engine.deployment
 
 import io.circe.{Decoder, Encoder, KeyDecoder, KeyEncoder}
+import net.ceedubs.ficus.Ficus
+import net.ceedubs.ficus.readers.ValueReader
 
 final case class EngineSetupName(value: String) {
   def withSuffix(suffix: String): EngineSetupName = EngineSetupName(value + suffix)
@@ -14,4 +16,6 @@ object EngineSetupName {
 
   implicit val keyEncoder: KeyEncoder[EngineSetupName] = KeyEncoder.encodeKeyString.contramap(_.value)
   implicit val keyDecoder: KeyDecoder[EngineSetupName] = KeyDecoder.decodeKeyString.map(EngineSetupName(_))
+
+  implicit val ficusReader: ValueReader[EngineSetupName] = Ficus.stringValueReader.map(EngineSetupName(_))
 }