+++ interpreter/src/test/scala/pl/touk/nussknacker/engine/compile/validationHelpers.scala
@@ -10,6 +10,7 @@ import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{CustomNod
 import pl.touk.nussknacker.engine.api.context.transformation._
 import pl.touk.nussknacker.engine.api.context.{ContextTransformation, JoinContextTransformation, ValidationContext}
 import pl.touk.nussknacker.engine.api.definition._
+import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.api.test.{TestData, TestRecord, TestRecordParser}
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, Unknown}
@@ -202,7 +203,7 @@ object validationHelpers {
     override def contextTransformation(context: ValidationContext, dependencies: List[NodeDependencyValue])(
         implicit nodeId: NodeId
     ): MissingParamHandleDynamicComponent$.ContextTransformationDefinition = { case TransformationStep(Nil, _) =>
-      NextParameters(Parameter[String]("param1") :: Nil)
+      NextParameters(Parameter[String](ParameterName("param1")) :: Nil)
     }
 
     override def implementation(
@@ -220,7 +221,7 @@ object validationHelpers {
     protected def outputParameters(
         context: ValidationContext,
         dependencies: List[NodeDependencyValue],
-        rest: List[(String, BaseDefinedParameter)]
+        rest: List[(ParameterName, BaseDefinedParameter)]
     )(implicit nodeId: NodeId): this.FinalResults = {
       dependencies.collectFirst { case OutputVariableNameValue(name) => name } match {
         case Some(name) =>
@@ -246,17 +247,17 @@ object validationHelpers {
       case TransformationStep(Nil, _) =>
         NextParameters(
           List(
-            Parameter("paramWithFixedValues", Typed[Int]).copy(editor =
+            Parameter(ParameterName("paramWithFixedValues"), Typed[Int]).copy(editor =
               Some(FixedValuesParameterEditor(List(FixedExpressionValue("1", "One"), FixedExpressionValue("2", "Two"))))
             )
           )
         )
       case TransformationStep(
-            ("paramWithFixedValues", DefinedEagerParameter(paramWithFixedValues: Int, _)) :: Nil,
+            (ParameterName("paramWithFixedValues"), DefinedEagerParameter(paramWithFixedValues: Int, _)) :: Nil,
             _
           ) =>
         FinalResults(context, state = Some(Valid(paramWithFixedValues)))
-      case TransformationStep(("paramWithFixedValues", FailedToDefineParameter) :: Nil, _) =>
+      case TransformationStep((ParameterName("paramWithFixedValues"), FailedToDefineParameter) :: Nil, _) =>
         FinalResults(context, state = Some(Invalid(())))
     }
 
@@ -274,7 +275,7 @@ object validationHelpers {
     protected def outputParameters(
         context: ValidationContext,
         dependencies: List[NodeDependencyValue],
-        rest: List[(String, BaseDefinedParameter)]
+        rest: List[(ParameterName, BaseDefinedParameter)]
     )(implicit nodeId: NodeId): this.FinalResults = {
       finalResult(context, rest, "otherNameThanInput")
     }
@@ -341,7 +342,7 @@ object validationHelpers {
 
         override def testParametersDefinition: List[Parameter] = Nil
 
-        override def parametersToTestData(params: Map[String, AnyRef]): String = ""
+        override def parametersToTestData(params: Map[ParameterName, AnyRef]): String = ""
       }
     }
 
@@ -352,7 +353,7 @@ object validationHelpers {
     protected def outputParameters(
         context: ValidationContext,
         dependencies: List[NodeDependencyValue],
-        rest: List[(String, BaseDefinedParameter)]
+        rest: List[(ParameterName, BaseDefinedParameter)]
     )(implicit nodeId: NodeId): this.FinalResults = {
       FinalResults(context)
     }
@@ -364,15 +365,16 @@ object validationHelpers {
     override def contextTransformation(context: ValidationContext, dependencies: List[NodeDependencyValue])(
         implicit nodeId: NodeId
     ): OptionalParametersSink.ContextTransformationDefinition = {
-      case TransformationStep(Nil, _) => NextParameters(List(Parameter.optional[CharSequence]("optionalParameter")))
-      case TransformationStep(("optionalParameter", _) :: Nil, None) =>
+      case TransformationStep(Nil, _) =>
+        NextParameters(List(Parameter.optional[CharSequence](ParameterName("optionalParameter"))))
+      case TransformationStep((ParameterName("optionalParameter"), _) :: Nil, None) =>
         outputParameters(context, dependencies, List())
     }
 
     protected def outputParameters(
         context: ValidationContext,
         dependencies: List[NodeDependencyValue],
-        rest: List[(String, BaseDefinedParameter)]
+        rest: List[(ParameterName, BaseDefinedParameter)]
     )(implicit nodeId: NodeId): this.FinalResults = {
       FinalResults(context)
     }
@@ -384,7 +386,7 @@ object validationHelpers {
     protected def outputParameters(
         context: ValidationContext,
         dependencies: List[NodeDependencyValue],
-        rest: List[(String, BaseDefinedParameter)]
+        rest: List[(ParameterName, BaseDefinedParameter)]
     )(implicit nodeId: NodeId): this.FinalResults = {
       FinalResults(context)
     }
@@ -398,7 +400,7 @@ object validationHelpers {
     protected def outputParameters(
         context: ValidationContext,
         dependencies: List[NodeDependencyValue],
-        rest: List[(String, BaseDefinedParameter)]
+        rest: List[(ParameterName, BaseDefinedParameter)]
     )(implicit nodeId: NodeId): this.FinalResults = {
       throw SomeException
     }
@@ -410,7 +412,7 @@ object validationHelpers {
     protected def outputParameters(
         context: ValidationContext,
         dependencies: List[NodeDependencyValue],
-        rest: List[(String, BaseDefinedParameter)]
+        rest: List[(ParameterName, BaseDefinedParameter)]
     )(implicit nodeId: NodeId): this.FinalResults = {
       dependencies.collectFirst { case OutputVariableNameValue(name) => name } match {
         case Some(name) =>
@@ -428,15 +430,27 @@ object validationHelpers {
 
     override type State = List[String]
 
+    private val par1ParamName     = ParameterName("par1")
+    private val lazyPar1ParamName = ParameterName("lazyPar1")
+
     override def contextTransformation(context: ValidationContext, dependencies: List[NodeDependencyValue])(
         implicit nodeId: NodeId
     ): this.ContextTransformationDefinition = {
       case TransformationStep(Nil, _) =>
-        NextParameters(List(Parameter[String]("par1"), Parameter[Long]("lazyPar1").copy(isLazyParameter = true)))
-      case TransformationStep(("par1", DefinedEagerParameter(value: String, _)) :: ("lazyPar1", _) :: Nil, None) =>
+        NextParameters(
+          List(
+            Parameter[String](par1ParamName),
+            Parameter[Long](lazyPar1ParamName).copy(isLazyParameter = true)
+          )
+        )
+      case TransformationStep(
+            (`par1ParamName`, DefinedEagerParameter(value: String, _)) :: (`lazyPar1ParamName`, _) :: Nil,
+            None
+          ) =>
         val split = value.split(",").toList
-        NextParameters(split.map(Parameter(_, Unknown)), state = Some(split))
-      case TransformationStep(("par1", _) :: ("lazyPar1", _) :: rest, Some(names)) if rest.map(_._1) == names =>
+        NextParameters(split.map(v => Parameter(ParameterName(v), Unknown)), state = Some(split))
+      case TransformationStep((`par1ParamName`, _) :: (`lazyPar1ParamName`, _) :: rest, Some(names))
+          if rest.map(_._1.value) == names =>
         outputParameters(context, dependencies, rest)
     }
 
@@ -446,9 +460,9 @@ object validationHelpers {
         outputVariable: Option[String]
     )(implicit nodeId: NodeId): FinalResults = {
       val result = Typed.record(
-        step.parameters.toMap.filterKeysNow(k => k != "par1" && k != "lazyPar1").map { case (k, v) =>
-          k -> v.returnType
-        }
+        step.parameters.toMap
+          .filterKeysNow(k => k != par1ParamName && k != lazyPar1ParamName)
+          .map { case (k, v) => k.value -> v.returnType }
       )
       prepareFinalResultWithOptionalVariable(inputContext, outputVariable.map(name => (name, result)), step.state)
     }
@@ -456,13 +470,17 @@ object validationHelpers {
     protected def outputParameters(
         context: ValidationContext,
         dependencies: List[NodeDependencyValue],
-        rest: List[(String, BaseDefinedParameter)]
+        rest: List[(ParameterName, BaseDefinedParameter)]
     )(implicit nodeId: NodeId): this.FinalResults
 
-    protected def finalResult(context: ValidationContext, rest: List[(String, BaseDefinedParameter)], name: String)(
+    protected def finalResult(
+        context: ValidationContext,
+        rest: List[(ParameterName, BaseDefinedParameter)],
+        name: String
+    )(
         implicit nodeId: NodeId
     ): this.FinalResults = {
-      val result = Typed.record(rest.map { case (k, v) => k -> v.returnType }.toMap)
+      val result = Typed.record(rest.map { case (k, v) => k.value -> v.returnType }.toMap)
       prepareFinalResultWithOptionalVariable(context, Some((name, result)), None)
     }
 
@@ -492,18 +510,21 @@ object validationHelpers {
       case TransformationStep(Nil, _) =>
         NextParameters(
           List(
-            Parameter("moreParams", Typed[Boolean]).copy(defaultValue = Some(Expression.spel("true")))
+            Parameter(ParameterName("moreParams"), Typed[Boolean]).copy(defaultValue = Some(Expression.spel("true")))
           )
         )
-      case TransformationStep(("moreParams", DefinedEagerParameter(true, _)) :: Nil, _) =>
+      case TransformationStep((ParameterName("moreParams"), DefinedEagerParameter(true, _)) :: Nil, _) =>
         NextParameters(
           List(
-            Parameter("extraParam", Typed[String])
+            Parameter(ParameterName("extraParam"), Typed[String])
               .copy(defaultValue = Some(Expression.spel(s"'$defaultExtraParamValue'")))
           )
         )
       case TransformationStep(
-            ("moreParams", _) :: ("extraParam", DefinedEagerParameter(extraParamValue: String, _)) :: Nil,
+            (ParameterName("moreParams"), _) :: (
+              ParameterName("extraParam"),
+              DefinedEagerParameter(extraParamValue: String, _)
+            ) :: Nil,
             _
           ) =>
         FinalResults(context, state = Some(extraParamValue))
@@ -527,18 +548,19 @@ object validationHelpers {
         contexts: Map[String, ValidationContext],
         dependencies: List[NodeDependencyValue]
     )(implicit nodeId: NodeId): DynamicParameterJoinTransformer.ContextTransformationDefinition = {
-      case TransformationStep(Nil, _) => NextParameters(List(Parameter[Boolean]("isLeft").copy(branchParam = true)))
+      case TransformationStep(Nil, _) =>
+        NextParameters(List(Parameter[Boolean](ParameterName("isLeft")).copy(branchParam = true)))
       case TransformationStep(
-            ("isLeft", DefinedEagerBranchParameter(byBranch: Map[String, Boolean] @unchecked, _)) :: Nil,
+            (ParameterName("isLeft"), DefinedEagerBranchParameter(byBranch: Map[String, Boolean] @unchecked, _)) :: Nil,
             _
           ) =>
         val error =
           if (byBranch.values.toList.sorted != List(false, true))
-            List(CustomNodeError("Has to be exactly one left and right", Some("isLeft")))
+            List(CustomNodeError("Has to be exactly one left and right", Some(ParameterName("isLeft"))))
           else Nil
         NextParameters(
           List(
-            Parameter[Any]("rightValue").copy(
+            Parameter[Any](ParameterName("rightValue")).copy(
               isLazyParameter = true,
               additionalVariables =
                 contexts(right(byBranch)).localVariables.mapValuesNow(AdditionalVariableProvidedInRuntime(_))
@@ -547,10 +569,8 @@ object validationHelpers {
           error
         )
       case TransformationStep(
-            ("isLeft", DefinedEagerBranchParameter(byBranch: Map[String, Boolean] @unchecked, _)) :: (
-              "rightValue",
-              rightValue: DefinedSingleParameter
-            ) :: Nil,
+            (ParameterName("isLeft"), DefinedEagerBranchParameter(byBranch: Map[String, Boolean] @unchecked, _)) ::
+            (ParameterName("rightValue"), rightValue: DefinedSingleParameter) :: Nil,
             _
           ) =>
         val out     = rightValue.returnType